core: add random color helper
